fix Shift entity
