+++ qr-shift-entity/src/entity/Shift.ts
@@ -1,0 +1,25 @@
+import z from "zod";
+
+const ShiftParser = z.object({
+  userID: z.string(),
+  startTime: z.date(),
+  endTime: z.date(),
+});
+
+export type ShiftRequest = z.infer<typeof ShiftParser>
+
+export const parseAsShiftRequest = (data: any): ShiftRequest | Error => {
+  const validationResponse = ShiftParser.safeParse(data);
+  return validationResponse.success
+    ? validationResponse.data
+    : validationResponse.error;
+};
+
+export type ConfirmedShift = z.infer<typeof ShiftParser>;
+
+export const parseAsConfirmedShift = (data: any): ConfirmedShift | Error => {
+  const validationResponse = ShiftParser.safeParse(data);
+  return validationResponse.success
+    ? validationResponse.data
+    : validationResponse.error;
+};
